Prepare an installable version of MetaPRL. The following directories are collected into libraries, and copied into a standard install location.
   libmojave
   mllib
   refiner
   filter
   support
   editor/ml

For each of these directories, the .mli,.cmi and the
library itself are copied to an <install>/lib directory.
In addition, the filter directory copies the camlp4{n,o}
executables.

Note that the other filter/ binaries are currently broken.
We'll need to fix someday.

The plan is:

   1. Install these files into a standard location like
      /usr/lib/metaprl.  TODO: add the location of the
      install directory to mk/config.

   2. Build a new OMakeroot that will allow people
      to build their own private copy of MetaPRL
      with their own theories, just from these
      binaries.

   3. (Probably) build a binary MetaPRL RPM.

Note: the current install target is called "export", because
we have a pre-existing "install" target.  I don't know what
the "install" was used for, but it seems like we should
usurp it for the current purpose.

I'll finish #2 tomorrow.